debug almost done, last step : put the enc28j60.cpp in hwfunk.cpp and write the functions

+++ ethercard/EtherCard.cpp
@@ -153,7 +153,7 @@ void Stash::prepare (const char* fmt, ...) {
     if (c == 0)
       break;
     if (c == '$') {
-      uint16_t argval = va_arg(ap, uint8_t), arglen = 0;
+      uint16_t argval = va_arg(ap,unsigned int), arglen = 0;
       switch (pgm_read_byte(fmt++)) {
         case 'D': {
           char buf[7];
